+++ llama_index/response_synthesizers/no_text.py
@@ -1,0 +1,21 @@
+from typing import Any, Sequence
+from llama_index.response_synthesizers.base import BaseSynthesizer
+from llama_index.types import RESPONSE_TEXT_TYPE
+
+
+class NoText(BaseSynthesizer):
+    def get_response(
+        self,
+        query_str: str,
+        text_chunks: Sequence[str],
+        **response_kwargs: Any,
+    ) -> RESPONSE_TEXT_TYPE:
+        return ""
+
+    async def aget_response(
+        self,
+        query_str: str,
+        text_chunks: Sequence[str],
+        **response_kwargs: Any,
+    ) -> RESPONSE_TEXT_TYPE:
+        return ""